+++ lib/presentation/weather/components/weather_temperature_display.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/svg.dart';
-import 'package:flutter_training/core/gen/assets.gen.dart';
 import 'package:flutter_training/domain/models/weather.dart';
+import 'package:flutter_training/presentation/weather/components/weather_image.dart';
 
 class WeatherTemperatureDisplay extends StatelessWidget {
   const WeatherTemperatureDisplay({
@@ -23,7 +22,7 @@ class WeatherTemperatureDisplay extends StatelessWidget {
       children: [
         AspectRatio(
           aspectRatio: 1,
-          child: _WeatherImage(
+          child: WeatherImage(
             weather: _weather,
           ),
         ),
@@ -48,33 +47,6 @@ class WeatherTemperatureDisplay extends StatelessWidget {
       ],
     );
   }
-}
-
-class _WeatherImage extends StatelessWidget {
-  const _WeatherImage({
-    required Weather? weather,
-  }) : _weather = weather;
-
-  final Weather? _weather;
-
-  @override
-  Widget build(BuildContext context) {
-    if (_weather == null) {
-      return const Placeholder();
-    } else {
-      return SvgPicture.asset(
-        _weather.assetName,
-      );
-    }
-  }
-}
-
-extension _WeatherAssetName on Weather {
-  String get assetName => switch (this) {
-        Weather.sunny => Assets.weather.sunny,
-        Weather.cloudy => Assets.weather.cloudy,
-        Weather.rainy => Assets.weather.rainy,
-      };
 }
 
 class _TemperatureLabel extends StatelessWidget {
